requirements.txt
certifi==2025.4.26
charset-normalizer==2.0.12
colour==0.1.5
dataclasses==0.8
docopt==0.6.2
idna==3.10
numpy==1.16.2
opencv-python==4.0.0.21
pandas==0.23.4
Pillow==5.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.27.1
six==1.17.0
torch==1.9.0
torchvision==0.10.0
typing_extensions==4.1.1
urllib3==1.26.20
yarg==0.1.10
